+++ src/components/Chat/AudioPlayer/index.tsx
@@ -1,6 +1,4 @@
-import React, { useState, useEffect, useCallback } from "react";
-import LottieView from "lottie-react-native";
-import { Audio } from "expo-av";
+import React, { useState } from "react";
 import {
   Container,
   AudioContainerWrapper,
@@ -14,7 +12,7 @@ import {
 import { MaterialIcons } from "@expo/vector-icons";
 import { useTheme } from "styled-components";
 import { millisToTime } from "../../../utils/format";
-import { useNavigation } from "@react-navigation/native";
+import { useFocusEffect  } from "@react-navigation/native";
 import { AudioData } from "../../../../@types/interfaces";
 import { useAudioPlayer } from "../../../contexts/audioPlayer";
 
@@ -24,69 +22,41 @@ interface IAudioPlayer {
 }
 
 const AudioPlayer = ({ audio, deleted }: IAudioPlayer) => {
-  const [isPlaying, setIsPlaying] = useState(false);
   const [currentPosition, setCurrentPosition] = useState(0);
-  const { loadAudio, playAndPauseAudio, currentAudioName } = useAudioPlayer();
+  const {
+    loadAudio,
+    playAndPauseAudio,
+    changeAudioPosition,
+    currentAudioName,
+  } = useAudioPlayer();
 
   const { colors } = useTheme();
-  const navigation = useNavigation();
 
-  useEffect(() => {
-    loadAudio({ name: audio.name, url: audio.url });
-  }, []);
+  useFocusEffect(() => {
+    loadAudio({
+      name: audio.name,
+      url: audio.url,
+      onStatusUpdate: async (status) => {
+        if (!status.isLoaded) return;
 
-  // useEffect(() => {
-  //   (async () => {
-  //     if (sound) {
-  //       if (isPlaying) {
-  //         await sound.pauseAsync();
-  //         setIsPlaying(false);
-  //       }
-  //       await sound.unloadAsync();
-  //     }
-  //   })();
-  // }, [deleted]);
-
-  // const loadAudio = useCallback(async () => {
-  //   if (sound) return;
-
-  //   const newSound = await Audio.Sound.createAsync({ uri: audio.url });
-
-  //   if (newSound.status.isLoaded) {
-  //     setSound(newSound.sound);
-  //     newSound.sound.setOnPlaybackStatusUpdate(async (status) => {
-  //       if (!status.isLoaded) return;
-
-  //       if (status.didJustFinish) {
-  //         await newSound.sound?.pauseAsync();
-  //         return await handleFinish();
-  //       } else {
-  //         setCurrentPosition(status.positionMillis);
-  //       }
-  //     });
-  //   }
-
-  //   navigation.addListener("blur", async () => {
-  //     if (newSound.sound._loaded) {
-  //       await newSound.sound.pauseAsync();
-  //       await newSound.sound.unloadAsync();
-  //       setIsPlaying(false)
-  //     }
-  //   });
-  // }, []);
-
-  async function handleFinish() {
-    setIsPlaying(false);
-    onChangePosition(0);
-  }
+        if (status.isPlaying) {
+          setCurrentPosition(status?.positionMillis);
+        }
+        
+      },
+      onFinishAudio: async () => {
+        setCurrentPosition(0);        
+      },
+    });
+  });
 
   const handlePlayPauseAudio = async () => {
     await playAndPauseAudio(audio.name, currentPosition);
   };
 
   const onChangePosition = async (value: number) => {
+    await changeAudioPosition(audio.name, value);
     setCurrentPosition(value);
-    // await sound?.setPositionAsync(value);
   };
 
   return (