+++ src/contexts/audioPlayer.tsx
@@ -3,9 +3,11 @@ import { Audio, AVPlaybackStatus } from "expo-av";
 
 interface AudioPlayerContextProps {
   loadAudio: (data: LoadAudioData) => Promise<void>;
-  playAndPauseAudio: (name: string, position: number) => void;
+  unloadAllAudios: () => Promise<void>;
+  playAndPauseAudio: (name: string, position?: number) => Promise<void>;
+  changeAudioPosition: (name: string, position: number) => Promise<void>;
   currentAudioName: string;
-  currentAudioData: AVPlaybackStatus;
+  currentSound: SoundData | null;
 }
 
 interface SoundData {
@@ -19,6 +21,8 @@ interface SoundData {
 interface LoadAudioData {
   name: string;
   url: string;
+  onStatusUpdate: (status: AVPlaybackStatus) => any;
+  onFinishAudio: () => any;
 }
 
 const AudioPlayerContext = createContext<AudioPlayerContextProps>(
@@ -28,26 +32,47 @@ const AudioPlayerContext = createContext<AudioPlayerContextProps>(
 const AudioPlayerProvider: React.FC = ({ children }) => {
   const [sounds, setSounds] = useState<SoundData[]>([]);
   const [currentAudioName, setCurrentAudioName] = useState("");
-  const [currentAudioData, setCurrentAudioData] = useState<AVPlaybackStatus>(
-    {} as AVPlaybackStatus
-  );
+  const [currentSound, setCurrentSound] = useState<SoundData | null>(null);
 
   useEffect(() => {
     Audio.setAudioModeAsync({
       shouldDuckAndroid: true,
-      allowsRecordingIOS: false,
+      allowsRecordingIOS: true,
       staysActiveInBackground: false,
       interruptionModeIOS: Audio.INTERRUPTION_MODE_IOS_DUCK_OTHERS,
       interruptionModeAndroid: Audio.INTERRUPTION_MODE_ANDROID_DUCK_OTHERS,
     });
   }, []);
 
-  const loadAudio = async ({ name, url }: LoadAudioData) => {
-    const hasSound = sounds.filter((sound) => sound.name === name);
+  const loadAudio = async ({
+    name,
+    url,
+    onStatusUpdate,
+    onFinishAudio,
+  }: LoadAudioData) => {
+    const hasSound = sounds.some((sound) => sound.name === name);
 
-    if (hasSound.length) return;
+    if (hasSound) return;
 
-    const audio = await Audio.Sound.createAsync({ uri: url });
+    const audio = await Audio.Sound.createAsync({ uri: url, name });
+
+    audio.sound.setOnPlaybackStatusUpdate(async (status) => {
+      if (!status.isLoaded) return;
+
+      if (status.didJustFinish) {        
+        await onFinishAudio();
+        await changeAudioPosition(name, 0)
+        await pauseAudio(name)
+
+        setCurrentAudioName("")
+        setCurrentSound(null)
+
+        return;
+      }
+
+      return await onStatusUpdate(status);
+    });
+
     setSounds((old) => [
       ...old,
       {
@@ -60,14 +85,33 @@ const AudioPlayerProvider: React.FC = ({ children }) => {
     ]);
   };
 
-  const playAndPauseAudio = async (name: string, position?: number) => {
-    let soundsForSave: SoundData[] = [];
+  const unloadAllAudios = async () => {
+    Promise.all(sounds.map(async s => {
+      if (!s) return;
+  
+      if (s.isPlaying) {
+        await s.controller.pauseAsync();
+      }
+  
+      if (s.name === currentAudioName) {
+        setCurrentAudioName("");
+        setCurrentSound(null);
+      }
+  
+      s.controller._clearSubscriptions()
+      await s.controller.unloadAsync();
+    }))
+
+    setSounds([]);
+  };
+
+  const playAudio = async (name: string, position = 0) => {
+    let soundsForSave = sounds.filter((s) => s.name !== name);
     const sound = sounds.filter((s) => s.name === name).shift();
-    soundsForSave = sounds.filter((s) => s.name !== name);
 
     if (!sound) return;
 
-    if (currentAudioName) {
+    if (currentAudioName !== sound.name) {
       soundsForSave = await Promise.all(
         soundsForSave.map(async (s) => {
           if (s.isPlaying) {
@@ -80,27 +124,53 @@ const AudioPlayerProvider: React.FC = ({ children }) => {
       );
     }
 
-    setCurrentAudioData(sound.data);
-    setCurrentAudioName(name);
+    setCurrentAudioName(sound.name);
+    setCurrentSound(sound);
+    await sound?.controller.playFromPositionAsync(position)
+    sound.isPlaying = true;
 
-    if (!sound.isPlaying) {
-      await sound?.controller.playFromPositionAsync(position || 0);
-      sound.isPlaying = true;
-    } else {
-      await sound.controller.pauseAsync();
-      sound.isPlaying = false;
-      setCurrentAudioName("");
-    }
     setSounds([...soundsForSave, sound]);
+  }
+
+  const pauseAudio = async (name: string) => {
+    const sound = sounds.filter((s) => s.name === name).shift();
+    const soundsForSave = sounds.filter((s) => s.name !== name);
+
+    if (!sound) return;
+
+    setCurrentAudioName("");
+    setCurrentSound(null);
+    await sound?.controller.pauseAsync()
+    sound.isPlaying = false;
+
+    setSounds([...soundsForSave, sound]);
+  }
+
+  const playAndPauseAudio = async (name: string, position = 0) => {
+    if (currentAudioName === name) {
+      await pauseAudio(name)
+    } else {
+      await playAudio(name, position)
+    }
+  };
+
+  const changeAudioPosition = async (name: string, position: number) => {
+    const sound = sounds.filter((s) => s.name === name).shift();
+
+    if (!sound) return;
+
+    await sound.controller.setPositionAsync(position);
   };
 
   return (
     <AudioPlayerContext.Provider
       value={{
         loadAudio,
+        unloadAllAudios,
         playAndPauseAudio,
+        changeAudioPosition,
         currentAudioName,
-        currentAudioData,
+        currentSound,
       }}
     >
       {children}